+++ example/di/property_module.dart
@@ -1,0 +1,10 @@
+import 'package:dependon/dependon.dart';
+
+import '../data/settings/settings.dart';
+
+class PropertyModule extends Module {
+  @override
+  void dependencies() {
+    property<Settings>(Settings());
+  }
+}