+++ example/main.dart
@@ -1,11 +1,13 @@
+import 'package:dependon/dependon.dart';
 import 'package:flutter/material.dart';
 
-import 'di/dependency_injection_module.dart';
+import 'di/repository_module.dart';
 import 'infrastructure/app.dart';
 
-void main() async {
-  final diModule = DependencyInjectionModule();
-  diModule.load();
+void main() {
+  final di = DependencyInjection.instance;
+  di.loadModule(RepositoryModule());
 
+  // etc
   runApp(myApp);
 }