+++ lib/src/enum/log_level.dart
@@ -1,0 +1,9 @@
+/// Level of logging.
+///
+/// [debug] = Everything will be printed.
+///
+/// [none] = No log will be printed.
+enum LogLevel {
+  debug,
+  none,
+}